Move selectors to separate files
