+++ ui/src/redux/selectors/mapFilter.ts
@@ -1,0 +1,3 @@
+import { RootState } from '../store';
+
+export const selectMapFilter = (state: RootState) => state.mapFilter;
