+++ ui/src/redux/selectors/mapRouteEditor.ts
@@ -1,0 +1,46 @@
+import { createSelector } from '@reduxjs/toolkit';
+import { mapFromStoreType } from '../mappers';
+import { MapRouteEditorState } from '../slices/mapRouteEditor';
+import { RootState } from '../store';
+
+export const selectMapRouteEditor = (state: RootState) =>
+  mapFromStoreType<MapRouteEditorState>(state.mapRouteEditor);
+
+export const selectIsInViewMode = createSelector(
+  selectMapRouteEditor,
+  (modalMap) => modalMap.drawingMode === undefined,
+);
+
+export const selectHasChangesInProgress = createSelector(
+  selectMapRouteEditor,
+  selectIsInViewMode,
+  (modalMap, viewMode) => modalMap.creatingNewRoute || !viewMode,
+);
+
+export const selectEditedRouteData = createSelector(
+  selectMapRouteEditor,
+  (mapRouteEditor) => mapRouteEditor.editedRouteData,
+);
+
+export const selectHasDraftRouteGeometry = createSelector(
+  selectEditedRouteData,
+  (editedRouteData) => !!editedRouteData.geometry,
+);
+
+export const selectSelectedRouteId = createSelector(
+  selectMapRouteEditor,
+  (mapRouteEditor) => mapRouteEditor.selectedRouteId,
+);
+
+export const selectDrawingMode = createSelector(
+  selectMapRouteEditor,
+  (mapRouteEditor) => mapRouteEditor.drawingMode,
+);
+
+export const selectEditedRouteIncludedStops = createSelector(
+  selectMapRouteEditor,
+  ({ editedRouteData }) =>
+    editedRouteData.stopsEligibleForJourneyPattern.filter((stop) =>
+      editedRouteData.includedStopLabels.includes(stop.label),
+    ),
+);
