+++ ui/src/redux/selectors/modals.ts
@@ -1,0 +1,29 @@
+import { createSelector } from '@reduxjs/toolkit';
+import { RootState } from '../store';
+
+export const selectModals = (state: RootState) => state.modals;
+
+export const selectViaModal = createSelector(
+  selectModals,
+  (modals) => modals.viaModal,
+);
+
+export const selectIsViaModalOpen = createSelector(
+  selectModals,
+  (modals) => modals.viaModal.isOpen,
+);
+
+export const selectIsTimingPlaceModalOpen = createSelector(
+  selectModals,
+  (modals) => modals.timingPlaceModal.isOpen,
+);
+
+export const selectTimingSettingsModal = createSelector(
+  selectModals,
+  (modals) => modals.timingSettingsModal,
+);
+
+export const selectIsTimingSettingsModalOpen = createSelector(
+  selectTimingSettingsModal,
+  (timingSettingsModal) => timingSettingsModal.isOpen,
+);
