+++ ui/src/redux/selectors/timetable.ts
@@ -1,0 +1,3 @@
+import { RootState } from '../store';
+
+export const selectTimetable = (state: RootState) => state.timetable;
